Move md5 file operations into vine_checksum, in preparation for some additions.
