+++ taskvine/src/manager/Makefile
@@ -16,6 +16,7 @@ SOURCES = \
 	vine_mount.c \
 	vine_txn_log.c \
 	vine_cached_name.c \
+	vine_checksum.c \
 	vine_perf_log.c \
 	vine_remote_file_info.c \
 	vine_factory_info.c \
